+++ classes/local/contexts/block_video_context.php
@@ -1,0 +1,113 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle. If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Defines a Moodle video file inside a block context.
+ *
+ * @package tool_openveo_migration
+ * @copyright 2019 Veo-labs
+ * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace tool_openveo_migration\local\contexts;
+
+defined('MOODLE_INTERNAL') || die();
+
+use stdClass;
+use context_block;
+use stored_file;
+use tool_openveo_migration\local\contexts\video_context;
+
+/**
+ * Defines a block context holding a Moodle video file.
+ *
+ * @package tool_openveo_migration
+ * @copyright 2019 Veo-labs
+ * @license http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class block_video_context extends video_context {
+
+    /**
+     * The block associated to the Moodle context.
+     *
+     * @var stdClass
+     */
+    protected $block;
+
+    /**
+     * The course associated to the block if any.
+     *
+     * @var stdClass
+     */
+    protected $course;
+
+    /**
+     * Builds a block context.
+     *
+     * @param stored_file $video The video that belongs to this context (either the original or an alias)
+     * @param context_block $moodlecontext The Moodle context
+     * @param stdClass $block The block associated to the Moodle context
+     * @param stdClass $course The course associated to the block if any
+     */
+    public function __construct(stored_file &$video, context_block &$moodlecontext, stdClass &$block,
+            stdClass &$course = null) {
+        parent::__construct($video, $moodlecontext);
+        $this->block = $block;
+        $this->course = $course;
+        $this->type = 'block';
+    }
+
+    /**
+     * Resolves a text regarding the context.
+     *
+     * Use this function to resolve a text by replacing its tokens by their values inside the context.
+     * Block context has the following available tokens:
+     * - %filename%: The Moodle video file name
+     * - %blockid%: The block id
+     * - %blockname%: The block name
+     *
+     * The following tokens are only available if the block has an associated course, if not tokens will be replaced by an empty
+     * string.
+     * - %courseid%: The course id
+     * - %courseidnumber%: The course id number
+     * - %coursecategoryid%: The id of the category the course belongs to
+     * - %coursefullname%: The course full name
+     * - %courseshortname%: The course short name
+     *
+     * @param string text The text containing tokens to resolve
+     * @return string The resolved text
+     */
+    public function resolve_text(string $text) : string {
+        return str_replace(array(
+            '%blockid%',
+            '%blockname%',
+            '%courseid%',
+            '%courseidnumber%',
+            '%coursecategoryid%',
+            '%coursefullname%',
+            '%courseshortname%'
+        ), array(
+            $this->block->id,
+            $this->block->name,
+            !empty($this->course) ? $this->course->id : '',
+            !empty($this->course) ? $this->course->idnumber : '',
+            !empty($this->course) ? $this->course->category : '',
+            !empty($this->course) ? $this->course->fullname : '',
+            !empty($this->course) ? $this->course->shortname : ''
+        ), parent::resolve_text($text));
+    }
+
+}